restructure meta-ai module
