+++ superai/apis/meta_ai/__init__.py
@@ -1,0 +1,2 @@
+from .model import *
+
